feat: Set semester value for new courses in COURSE_MAPPING
When a new course is discovered during file upload, the `semester` attribute in the `COURSE_MAPPING` table is now populated based on the semester in which the user took the course.

- 1st semester courses are set to `1`.
- 2nd semester courses are set to `2`.
- Summer/Winter semester courses are set to `3`.

This prevents the semester from being `null` for newly added courses and allows for better data integrity and future filtering capabilities.