refactor: Set semester to 4 for all new courses
Modified the logic for handling new courses found during file upload.
Instead of determining the semester based on when the user took the course, all newly discovered courses will now have their `semester` attribute in the `COURSE_MAPPING` table set to a default value of `4`. This simplifies the logic as requested.